better cache preference and google invoking logic
First, if the cache matches something but it scores
it really low so as to not show up, then don't tell the google
search logic to skip over that entry because the IMDB id was
already "seen"

Second, actually just score the cache near 100%. Calculate what
the penalty WOULD have been (but leave the cache score high).  Use
the penalty to determine if google should be consulted. This will
allow FIM to consult google more frequently if the user is indeed
wanting to correct the cache hit.