scoring tweaks and substring bug fix
the longest substring match was not float()-ing the entire
fraction causing the logic to fail except in the case of '0' character
match (which is very rare since usually at least one vowel would
match.

also, the cache items are given a score based on the score penalty
but since we trust the cache, we limit it to a lower bound of 85
(the cache cut off).  This allows cached items to compete with
one another but not dissapear from the results